Modularied Search Bar Logic
