+++ src/App.js
@@ -1,17 +1,13 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import "./App.css";
 import SearchBar from "./components/input/SearchBar";
 import logo from "./mlh-prep.png";
-import { getAutocompleteCities } from "./utils";
 
 function App() {
-  const inputRef = useRef();
   const [error, setError] = useState(null);
   const [isLoaded, setIsLoaded] = useState(false);
   const [city, setCity] = useState("New York City");
   const [results, setResults] = useState(null);
-  const [autocompleteCities, setAutocompleteCities] = useState([]);
-  let reqTimeout = null;
 
   useEffect(() => {
     fetch(
@@ -34,24 +30,6 @@ function App() {
       );
   }, [city]);
 
-  const onInputChange = async (e) => {
-    if (reqTimeout) {
-      clearTimeout(reqTimeout);
-    }
-
-    const cityValue = inputRef.current.value;
-
-    reqTimeout = getAutocompleteCities(cityValue, setAutocompleteCities);
-  };
-
-  const handleKey = async (e) => {
-    const cityValue = inputRef.current.value;
-
-    if (e.code === "Enter") {
-      setCity(cityValue);
-    }
-  };
-
   if (error) {
     return <div>Error: {error.message}</div>;
   }
@@ -62,16 +40,11 @@ function App() {
       <div>
         <h2>Enter a city below 👇</h2>
 
-        <SearchBar
-          inputRef={inputRef}
-          onKeyDown={handleKey}
-          onChange={onInputChange}
-          autocompleteCities={autocompleteCities}
-        />
+        <SearchBar setCity={setCity} />
 
         <div className="Results">
           {!isLoaded && <h2>Loading...</h2>}
-          {console.log(results)}
+          {/* {console.log(results)} */}
           {isLoaded && results && (
             <>
               <h3>{results.weather[0].main}</h3>
